Rename defaultinitalloc.hpp to defaultinitvec.hpp
