+++ src/gfx/reverse.cpp
@@ -12,7 +12,7 @@
 #include <string.h>
 #include <vector>
 
-#include "defaultinitalloc.hpp"
+#include "defaultinitvec.hpp"
 #include "file.hpp"
 #include "helpers.hpp" // assume
